requirements.txt
flake8==3.9.2
flake8-docstrings==1.6.0
pytest==6.2.5
python-dotenv==0.19.0
aiogram==2.23.1
requests==2.26.0
asyncio==3.4.3
ngrok==0.1.6
